+++ Source/UI/Meterbridge.cpp
@@ -2,6 +2,7 @@
 #include "../Theme/BrandColors.h"
 #include "../Theme/BrandTokens.h"
 #include "LedMeter.h"
+#include "FloatingLaunch.h"
 
 namespace zynforge
 {
@@ -77,6 +78,6 @@ namespace zynforge
         opts.escapeKeyTriggersCloseButton = true;
         opts.useNativeTitleBar            = true;
         opts.resizable                    = true;
-        return opts.launchAsync();
+        return launchFloating (opts);   // non-modal so the METERS tab can toggle it
     }
 }